+++ webpack.config.js
@@ -24,6 +24,10 @@ module.exports = {
           { loader: 'css-loader' },
         ],
       },
+      {
+        test: /\.svg$/,
+        loader: 'svg-inline-loader'
+      },
     ],
   },
   plugins: [